BigDecimal em valores

+++ src/main/java/com/cliente/recepcaopedido/interfaces/repository/PedidoRepository.java
@@ -1,6 +1,7 @@
 package com.cliente.recepcaopedido.interfaces.repository;
 
 import com.cliente.recepcaopedido.model.PedidoModel;
+import java.math.BigDecimal;
 import java.util.List;
 import org.springframework.data.repository.CrudRepository;
 
@@ -12,5 +13,5 @@ public interface PedidoRepository extends CrudRepository<PedidoModel, Integer>{
     public List<PedidoModel> findAllByDataCadastro(String data);
     public PedidoModel findByNumeroControle(int numeroControle);
     public List<PedidoModel> findAllByCodigoCliente(int codigoCliente);
-    public List<PedidoModel> findAllByValorTotal(float valorTotal);
+    public List<PedidoModel> findAllByValorTotal(BigDecimal valorTotal);
 }
